+++ dataforge-context/src/jvmTest/kotlin/space/kscience/dataforge/descriptors/TestAutoDescriptors.kt
@@ -1,3 +1,5 @@
+@file:OptIn(DFExperimental::class)
+
 package space.kscience.dataforge.descriptors
 
 import kotlinx.serialization.encodeToString
@@ -8,8 +10,9 @@ import space.kscience.dataforge.meta.SchemeSpec
 import space.kscience.dataforge.meta.descriptors.MetaDescriptor
 import space.kscience.dataforge.meta.int
 import space.kscience.dataforge.meta.string
+import space.kscience.dataforge.misc.DFExperimental
 
-private class TestScheme: Scheme(){
+private class TestScheme : Scheme() {
 
     @Description("A")
     val a by string()
@@ -17,15 +20,17 @@ private class TestScheme: Scheme(){
     @Description("B")
     val b by int()
 
-    companion object: SchemeSpec<TestScheme>(::TestScheme){
+    val c by int()
+
+    companion object : SchemeSpec<TestScheme>(::TestScheme) {
         override val descriptor: MetaDescriptor = autoDescriptor()
     }
 }
 
 class TestAutoDescriptors {
     @Test
-    fun autoDescriptor(){
+    fun autoDescriptor() {
         val autoDescriptor = MetaDescriptor.forClass(TestScheme::class)
-        println(Json{prettyPrint = true}.encodeToString(autoDescriptor))
+        println(Json { prettyPrint = true }.encodeToString(autoDescriptor))
     }
 }